@@ -74,11 +74,11 @@
     "sinon": "^1.17.3",
     "sinon-chai": "^2.8.0",
     "url-loader": "^0.5.7",
-    "vue": "^2.1.0",
-    "vue-loader": "^10.0.0",
-    "vue-router": "^2.1.1",
-    "vue-style-loader": "^1.0.0",
-    "vue-template-compiler": "^2.1.0",
+    "vue": "^2.2.1",
+    "vue-loader": "^11.1.4",
+    "vue-router": "^2.3.0",
+    "vue-style-loader": "^2.0.3",
+    "vue-template-compiler": "^2.2.1",
     "webpack": "^1.13.2",
     "webpack-dev-middleware": "^1.8.3",
     "webpack-hot-middleware": "^2.12.2",
